requirements.txt
aws-wsgi==0.2.7
bcrypt==4.1.3
blinker==1.8.2
boto3==1.34.103
botocore==1.34.103
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-PyMongo==2.3.0
fpdf2==2.8.3
gunicorn==20.1.0
httptools==0.6.1
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
marshmallow==3.26.1
joblib==1.4.2
openpyxl==3.1.5
packaging==24.0
pandas==2.2.3
PyJWT==2.8.0
Pillow==11.1.0
pymongo==4.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
urllib3==1.26.18
websockets==13.0
Werkzeug==3.0.3
zipp==3.18.1